vectordb/requirements.txt
faiss-cpu==1.8.0
Flask==3.0.3
numpy==1.26.4
setproctitle==1.3.3
packaging==24.1
openpyxl==3.1.5
pandas==2.2.3